requirements.txt
alembic==0.8.7
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
click==6.6
dominate==2.2.1
et-xmlfile==1.0.1
eventlet==0.19.0
Flask==0.11.1
Flask-Assets==0.11
Flask-Bootstrap==3.3.6.0
Flask-Login==0.3.2
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SocketIO==2.6
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
greenlet==0.4.10
gunicorn==19.6.0
itsdangerous==0.24
jdcal==1.2
Jinja2==2.8
kombu==3.0.35
Mako==1.0.4
MarkupSafe==0.23
openpyxl==2.3.5
progressbar2==3.10.0
python-editor==1.0.1
python-engineio==0.9.2
python-socketio==1.4.3
python-utils==2.0.0
pytz==2016.6.1
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.14
visitor==0.1.3
webassets==0.11.1
Werkzeug==0.11.10
WTForms==2.1
